Updated files and changelog for v0.9.20
