+++ ChangeLog
@@ -1,3 +1,34 @@
+k2hdkc (0.9.20) unstable; urgency=low
+
+  * Update dependency on k2hash and chmpx
+
+ -- Takeshi Nakatani <ggtakec@gmail.com>  Tue, 27 Mar 2018 13:09:22 +0900
+
+k2hdkc (0.9.19) unstable; urgency=low
+
+  * Bump up version with nothing update
+
+ -- Hirotaka Wakabayashi <hiwakaba@yahoo-corp.jp>  Mon, 26 Feb 2018 09:58:21 +0900
+
+k2hdkc (0.9.18) unstable; urgency=low
+
+  * Bump up version with nothing update
+
+ -- Hirotaka Wakabayashi <hiwakaba@yahoo-corp.jp>  Wed, 21 Feb 2018 16:15:46 +0900
+
+k2hdkc (0.9.17) unstable; urgency=low
+
+  * Fixed C++ statements mixed in lib/k2hdkccomstructure.h
+
+ -- Hirotaka Wakabayashi <hiwakaba@yahoo-corp.jp>  Mon, 19 Feb 2018 13:56:42 +0900
+
+k2hdkc (0.9.16) unstable; urgency=low
+
+  * Fixed C++ statements mix in including k2hdkc.h
+  * update readme
+
+ -- Takeshi Nakatani <ggtakec@gmail.com>  Wed, 14 Feb 2018 18:24:31 +0900
+
 k2hdkc (0.9.15) unstable; urgency=low
 
   * First version of open sorce on Github
